+++ PoliticPolls.Web/Controllers/PollsController.cs
@@ -23,7 +23,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Polls/Details/5
-        public ActionResult Details(int? id)
+        public ActionResult Details(decimal? id)
         {
             if (id == null)
             {
@@ -63,7 +63,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Polls/Edit/5
-        public ActionResult Edit(int? id)
+        public ActionResult Edit(decimal? id)
         {
             if (id == null)
             {
@@ -96,7 +96,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Polls/Delete/5
-        public ActionResult Delete(int? id)
+        public ActionResult Delete(decimal? id)
         {
             if (id == null)
             {
@@ -113,7 +113,7 @@ namespace PoliticPolls.Web.Controllers
         // POST: Polls/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(decimal id)
         {
             var poll = db.Poll.Find(id);
             db.Poll.Remove(poll);
