+++ PoliticPolls.Web/Controllers/RespondentsController.cs
@@ -21,7 +21,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Respondents/Details/5
-        public ActionResult Details(int? id)
+        public ActionResult Details(decimal? id)
         {
             if (id == null)
             {
@@ -59,7 +59,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Respondents/Edit/5
-        public ActionResult Edit(int? id)
+        public ActionResult Edit(decimal? id)
         {
             if (id == null)
             {
@@ -90,7 +90,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Respondents/Delete/5
-        public ActionResult Delete(int? id)
+        public ActionResult Delete(decimal? id)
         {
             if (id == null)
             {
@@ -107,7 +107,7 @@ namespace PoliticPolls.Web.Controllers
         // POST: Respondents/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(decimal id)
         {
             var respondents = db.Respondents.Find(id);
             db.Respondents.Remove(respondents);
